added function getSchemabyMessage in order to get parsed schema from the schema ID
I need to parse once and batch handle thousands of messages, and parsing the schema and decoding at the same time proved to be inefficient. Adding this function allows me to grab the schema once by ID and then iterate over messages to decode it. This was the quickest method I came up with, however I think your internal schema cache should be extended to cache the already parsed schema